Changes to hashing that solve the xinetd boolean benchmark in 14s (from ~25min). Switched to standard hash_set, hash_map, new hash for the vector of node values (from boost), changed the hash for nodes to be over id's, all the hash values are now size_t. The parser is down from 11s to 10s on the benchmark, so most of the solve time is parsing and we need to figure this out.

+++ src/expr/expr.cpp
@@ -77,7 +77,7 @@ bool Expr::operator<(const Expr& e) const {
   return *d_node < *e.d_node;
 }
 
-uint64_t Expr::hash() const {
+size_t Expr::hash() const {
   Assert(d_node != NULL, "Unexpected NULL expression pointer!");
   return (d_node->isNull());
 }
